In order to have a config setting that is toggleable, we need to modify a static file that contains the property. I tried to achieve this with a config.js file, but as soon as the context window closes, the toggle goes back to whatever value it had before.
As a reuslt, I tried creating a config JSON so that config settings are modifyable using the command line interface. This backfired as import type assertions for JSON files are not working in my environment, and I dont want to risk this being broken for other people in their deployments.

As a result, the fucntion is always enabled, but the config.js file will read env variables for text logging instead.